feat(hkl3d slicer): The HKL_3D_Slicer standalone mode
HKL 3D Slicer standalone mode allows for data to be loaded in rather
than needing parent data to load

add: toggle for the min and max intensity
add(color map): jet, virdis
remove(color map): spring, cool
fix(gui): add the original resolution to the gui

fix(area_det): size manager by adding path to go to the parent directory
'DashPVA'

add(dashpva cli): sim command to start the pva workflow setup
add(dashpva cli): slice3d command to open the HKL3d in standalone mode
